pyproject.toml
[tool.poetry]
name = "e-commerce-back-django"
version = "0.1.0"
description = ""
authors = ["Antonio Cortez <antoniocortez1497@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">4.2 <5"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
django-simple-history = "^3.7.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
